[BUGFIX] Fix return type of GU::getMaxUploadFileSize
This method promises to return an int, but actually returns a float.

Also add a regression test.

> ./Build/Scripts/runTests.sh -s phpstanGenerateBaseline

Resolves: #101499
Releases: main, 12.4, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80273
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/core/Classes/Utility/GeneralUtility.php
@@ -2231,7 +2231,7 @@ class GeneralUtility
      * This might be handy to find out the real upload limit that is possible for this
      * TYPO3 installation.
      *
-     * @return int The maximum size of uploads that are allowed (measured in kilobytes)
+     * @return int Maximum size of uploads that are allowed in KiB (divider 1024)
      */
     public static function getMaxUploadFileSize()
     {
@@ -2244,7 +2244,7 @@ class GeneralUtility
         // If the total amount of post data is smaller (!) than the upload_max_filesize directive,
         // then this is the real limit in PHP
         $phpUploadLimit = $phpPostLimit > 0 && $phpPostLimit < $phpUploadLimit ? $phpPostLimit : $phpUploadLimit;
-        return floor($phpUploadLimit) / 1024;
+        return (int)(floor($phpUploadLimit) / 1024);
     }
 
     /**
